[38635] Finished connecting the dots between an incoming packet from the Tizen server all the way up to the MainActivity. Fixed some of the bugs. Need to test thoroughly and fix whatever bugs there are in the json parsing code. Need to continue testing all code above that. Need to implement UI changes based on service updates. Need to refactor service object to be local/remote agnostic. Need to unify UI behavior between all HVAC apps, and add remaining services to Android app.

+++ app/src/main/java/com/jaguarlandrover/hvacdemo/RVIDlinkAuthPacket.java
@@ -55,7 +55,7 @@ public class RVIDlinkAuthPacket extends RVIDlinkPacket
         super(Command.AUTHORIZE, jsonHash);
 
         mAddr = (String)  jsonHash.get("addr");
-        mPort = (Integer) jsonHash.get("port");
+        mPort = ((Double) jsonHash.get("port")).intValue();
         mVer  = (String)  jsonHash.get("ver");
         mCert = (String)  jsonHash.get("cert");
     }
